Avoid caching jaas-dashboard clone
The jaas-dashboard clone was being cached in subsequent builds, so we
weren't pulling in the latest version of the repo for every build.

This should fix that. @jkfran found
[this technique](https://stackoverflow.com/questions/36996046/how-to-prevent-dockerfile-caching-git-clone)
of using `ADD` to query https://api.github.com/,
but unfortunately it didn't work for me because I kept getting the error
"invalid not-modified ETag" from GitHub. This
[should have been fixed](https://github.com/moby/buildkit/issues/905)
but the fix doesn't seem to have made its way into latest docker-ce yet.
So at some point in the future we could use this solution.

For the time being I've instead used https://httpbin.org/uuid to create
a cache-busting line in the `Dockerfile` just above the git clone.
This is less efficient, in that the GitHub solution would allow
GitHub to return a 304 if nothing had changed, which would allow
Docker to make use of the further caching below that point, whereas in
this implementation it will simply clone the repo every time. But I
believe this is the best we can do for now.